Introduce another tracking of state, context.
Refactor keybindings into a map that we can run all at once. Modifying
and tinkering with bindings is now much nicer because we clear all
previous crap away at load.

Showing the quick tag popup is now tied to a (:contexts db) set and
this same set is also being used to check if we want particular key
bindings to be active. I don't think this system will be the final one
but it's an idea for now. For one, it currently doesn't do a good job
allowing binding reuse across different contexts, which is frankly one
of the ideas behind maps and submaps.